packages/frontend/src/__tests__/task-form.test.tsx
import { vi } from 'vitest';
import { render, screen, fireEvent, waitFor } from '@testing-library/react';
import { Provider } from 'react-redux';
import { configureStore } from '@reduxjs/toolkit';
import { api } from '@/store/services/api';
import TaskForm from '@/components/task-form';

vi.mock('@/store/services/api', () => ({
  api: {
    reducerPath: 'api',
    reducer: (state = {}) => state,
    middleware: () => [],
  },
  useCreateTaskMutation: () => [mockCreateTask],
}));

const mockCreateTask = vi.fn();

const store = configureStore({
  reducer: {
    [api.reducerPath]: api.reducer,
  },
  middleware: (getDefaultMiddleware) =>
    getDefaultMiddleware().concat(api.middleware),
});

describe('TaskForm', () => {
  beforeEach(() => {
    mockCreateTask.mockClear();
  });

  it('renders form fields correctly', () => {
    render(
      <Provider store={store}>
        <TaskForm />
      </Provider>
    );

    expect(screen.getByLabelText(/task name/i)).toBeInTheDocument();
    expect(screen.getByLabelText(/description/i)).toBeInTheDocument();
    expect(screen.getByRole('button', { name: /create task/i })).toBeInTheDocument();
  });

  it('validates required fields', async () => {
    render(
      <Provider store={store}>
        <TaskForm />
      </Provider>
    );

    fireEvent.click(screen.getByRole('button', { name: /create task/i }));

    await waitFor(() => {
      expect(screen.getByText(/title must be at least 3 characters/i)).toBeInTheDocument();
      expect(screen.getByText(/description must be at least 10 characters/i)).toBeInTheDocument();
    });
  });

  it('submits form with valid data', async () => {
    render(
      <Provider store={store}>
        <TaskForm />
      </Provider>
    );

    fireEvent.change(screen.getByLabelText(/task name/i), {
      target: { value: 'Test Task' },
    });
    fireEvent.change(screen.getByLabelText(/description/i), {
      target: { value: 'Test Description for the task' },
    });

    fireEvent.click(screen.getByRole('button', { name: /create task/i }));

    await waitFor(() => {
      expect(mockCreateTask).toHaveBeenCalledWith({
        title: 'Test Task',
        description: 'Test Description for the task',
      });
    });
  });
});